Add NGitLab git submodule
This repository for now depends on changes in NGitLab not yet merged into the NGitLab repository or available as NuGet package.
To be able to use the required changes, add git submodule with the changes applied.
The customized version of NGitLab is the "integration" branch in https://github.com/ap0llo/NGitLab