Add Cake.GitLab.Testing package
Add package Cake.GitLab.Testing that provides a Fake/Mock implementation of IGitLabProvider to ease unit testing.

FakeGitLabProvider is automatically generated based on the IGitLabProvider interface (implemented via a source generator in Cake.GitLab.Generator).
For each method of IGitLabProvider, FakeGitLabProvider has a property named the same as the method that provides functionatliy to set up a delegate to call when the corresponding method is called and provides a log of all of the method's invovations.